Extract components inside DataMap to separate files
